+++ app/styles/components/modals.css
@@ -240,6 +240,8 @@
     display: flex;
     padding: 0;
     flex-grow: 1;
+}
+.labs-newsletter-settings .modal-fullsettings-body {
     overflow: hidden;
 }
 
@@ -265,6 +267,9 @@
     display: flex;
     flex-direction: column;
     padding: 0px 24px 20px;
+    width: 362px;
+}
+.labs-newsletter-settings .modal-fullsettings-sidebar {
     width: 400px;
     overflow-y: auto;
 }
@@ -276,17 +281,21 @@
 .modal-fullsettings-topbar {
     height: 66px;
     padding: 0 24px;
+}
+.labs-newsletter-settings .modal-fullsettings-topbar {
     border-bottom: 1px solid var(--whitegrey);
 }
 
 .modal-fullsettings-heading {
     display: flex;
-    height: 66px;
     align-items: center;
     font-size: 1.9rem;
     font-weight: 600;
     padding: 0 24px;
     margin: 0 -24px 1px;
+}
+.labs-newsletter-settings .modal-fullsettings-heading {
+    height: 66px;
 }
 
 .modal-fullsettings-form {
@@ -298,12 +307,20 @@
     padding: 0 24px;
 }
 
-/* .modal-fullsettings-section.first {
+/* labs-newsletter-settings - delete all when removing flag */
+.modal-fullsettings-section.first {
+    margin-top: 8px;
+}
+.labs-newsletter-settings .modal-fullsettings-section.first {
     margin-top: 24px;
-} */
+}
+/* end delete all */
 
 .modal-fullsettings-section.divider-top {
     border-top: 1px solid var(--whitegrey);
+    padding-top: 16px;
+}
+.labs-newsletter-settings .modal-fullsettings-section.divider-top {
     padding-top: 24px;
 }
 
@@ -321,6 +338,9 @@
     justify-content: space-between;
     align-items: center;
     padding: 0;
+    margin-bottom: 0;
+}
+.labs-newsletter-settings .modal-fullsettings-section .form-group {
     margin-bottom: 20px;
 }
 
@@ -369,7 +389,7 @@
     margin: 0 12px 0 0;
 }
 
-.modal-fullsettings-uploader {
+.labs-newsletter-settings .modal-fullsettings-uploader {
     display: flex;
     align-items: center;
     margin: 18px 0 0;
@@ -378,9 +398,9 @@
     border-radius: 3px;
 }
 
-.gh-header-img-uploadicon,
-.gh-header-img-uploadicon:hover,
-.gh-header-img-uploadicon:focus {
+.labs-newsletter-settings .gh-header-img-uploadicon,
+.labs-newsletter-settings .gh-header-img-uploadicon:hover,
+.labs-newsletter-settings .gh-header-img-uploadicon:focus {
     width: 56px;
     height: 44px;
     margin-right: 11px;
@@ -389,27 +409,27 @@
     box-shadow: none;
 }
 
-.gh-header-img-uploadicon span {
+.labs-newsletter-settings .gh-header-img-uploadicon span {
     display: flex;
     justify-content: center;
     align-items: center;
 }
 
-.gh-header-img-uploadicon span svg {
+.labs-newsletter-settings .gh-header-img-uploadicon span svg {
     width: 18px;
     height: 18px;
     fill: var(--darkgrey);
 }
 
-.gh-header-img-uploadicon:hover span svg {
+.labs-newsletter-settings .gh-header-img-uploadicon:hover span svg {
     fill: var(--darkgrey);
 }
 
-.gh-header-img-container {
+.labs-newsletter-settings .gh-header-img-container {
     height: 44px;
 }
 
-.gh-header-img-thumbnail {
+.labs-newsletter-settings .gh-header-img-thumbnail {
     display: inline-block;
     width: 56px;
     height: 44px;
@@ -421,11 +441,11 @@
     border-radius: 3px;
 }
 
-.gh-header-img-thumbnail svg path {
+.labs-newsletter-settings .gh-header-img-thumbnail svg path {
     stroke: var(--midlightgrey-d1);
 }
 
-.gh-header-img-deleteicon {
+.labs-newsletter-settings .gh-header-img-deleteicon {
     position: absolute;
     left: 4px;
     width: 56px;
@@ -434,17 +454,17 @@
     opacity: 0;
 }
 
-.gh-header-img-deleteicon:hover {
+.labs-newsletter-settings .gh-header-img-deleteicon:hover {
     opacity: 1;
 }
 
-.gh-header-img-deleteicon span {
+.labs-newsletter-settings .gh-header-img-deleteicon span {
     display: flex;
     justify-content: center;
     align-items: center;
 }
 
-.gh-header-img-deleteicon span svg {
+.labs-newsletter-settings .gh-header-img-deleteicon span svg {
     width: 18px;
     height: 18px;
 }
@@ -473,16 +493,22 @@
 .modal-fullsettings-preview-container {
     overflow: hidden;
     background: var(--whitegrey-l1);
-    height: 100vh;
+    height: calc(100vh - 126px);
     overflow-y: scroll;
+}
+.labs-newsletter-settings .modal-fullsettings-preview-container {
+    height: 100vh;
 }
 
 .modal-fullsettings-preview-hidescrollbar {
     overflow: hidden;
-    height: 100vh;
+    height: calc(100vh - 126px);
     background: var(--whitegrey-l1);
     border: 1px solid var(--whitegrey);
     border-radius: 5px;
+}
+.labs-newsletter-settings .modal-fullsettings-preview-hidescrollbar {
+    height: 100vh;
 }
 
 .modal-fullsettings-preview-hidescrollbar .modal-fullsettings-preview-container {
