+++ app/templates/settings/members-email.hbs
@@ -45,8 +45,14 @@
 
 {{#if this.showEmailDesignSettings}}
     <GhFullscreenModal @modifier="full-overlay portal-settings">
-        <ModalEmailDesignSettings
-            @closeModal={{this.toggleEmailDesignSettings}}
-        />
+        {{#if (enable-developer-experiments)}}
+            <ModalEmailDesignSettingsLabs
+                @closeModal={{this.toggleEmailDesignSettings}}
+            />
+        {{else}}
+            <ModalEmailDesignSettings
+                @closeModal={{this.toggleEmailDesignSettings}}
+            />
+        {{/if}}
     </GhFullscreenModal>
 {{/if}}
